requirements.txt
asgiref==3.2.10
async-lru==1.0.2
Django==3.1.1
djangorestframework==3.11.1
environs==8.0.0
marshmallow==3.7.1
psycopg2-binary==2.8.6
pyaes==1.6.1
Pyrogram==1.0.7
PySocks==1.7.1
python-dotenv==0.14.0
pytz==2020.1
sqlparse==0.3.1
